[backend] More Documentation
Separated the Calendar routes between protected and public as getting the ics is meant to be for students. And some bug fixes for translations
